requirements.txt
streamlit 
langchain 
langchain-groq 
langchain-huggingface 
faiss-cpu 
pandas 
python-dotenv